requirements.txt
discord.py[voice, speed]@git+https://github.com/Rapptz/discord.py.git@master
GitPython  # Optional
